+++ example/reload_module/client.py
@@ -11,7 +11,8 @@ async def sync_sum(a: int, b: int) -> int:
 
 
 async def main() -> None:
-    await client.connect()
+    client.add_conn("localhost", 9000)
+    await client.start()
     print(f"sync result: {await client.call(sync_sum, [1, 2])}")
     print(f"reload :{ await client.raw_call('reload', ['test_module', 'sync_sum'], group='registry')}")
     print(f"sync result: {await client.raw_call('sync_sum', [1, 2])}")